+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Add case for a template resource_type in append_data method in PaginatedChildrenResult model
+- Add `Proof` model and `ProofType` enum
+- Add a proof field to the `Row` model
+- Add serialization tests for the new model
 
 ## [4.2.0] - 2026-07-09